Fix PR 110539: missed optimization after moving two_value to match.pd
When I moved two_value to match.pd, I removed the check for the {0,+-1}
as I had placed it after the {0,+-1} case for cond in match.pd.
In the case of {0,+-1} and non boolean, before we would optmize those
case to just `(convert)a` but after we would get `(convert)(a != 0)`
which was not handled anyways to just `(convert)a`.
So this adds a pattern to match `(convert)(zeroone != 0)` and simplify
to `(convert)zeroone`.

In the bug report, we do finally optimize `(convert)(zeroone != 0)` to
`zeroone` in VRP2 but that in itself is too late and we miss other
optimizations that would have happened.

OK? Bootstrapped and tested on x86_64-linux-gnu with no regressions.

gcc/ChangeLog:

	PR tree-optimization/110539
	* match.pd ((convert)(zeroone !=/== 0)): Match
	and simplify to ((convert)zeroone)){,^1}.

gcc/testsuite/ChangeLog:

	PR tree-optimization/110539
	* gcc.dg/tree-ssa/pr110539-1.c: New test.
	* gcc.dg/tree-ssa/pr110539-2.c: New test.
	* gcc.dg/tree-ssa/pr110539-3.c: New test.