genopinit: Allow more than 256 modes.
Hi,

upcoming changes for RISC-V will have us exceed 256 modes or 8 bits. The
helper functions in gen* rely on the opcode as well as two modes fitting
into an unsigned int (a signed int even if we consider the qsort default
comparison function).  This patch changes the type of the index/hash
from unsigned int to unsigned long long and allows up to 16 bits for a
mode as well as 32 bits for an optab.

Despite fearing worse, bootstrap, build and test suite run times on
x86, aarch64, rv64 and power10 are actually unchanged (I didn't check
32-bit architectures but would expect similar results).

Regards
 Robin

gcc/ChangeLog:

	* genopinit.cc (pattern_cmp): Use if/else for comparison instead
	of subtraction.
	(main): Change to unsigned long long.
	* gensupport.cc (find_optab): Ditto.
	* gensupport.h (struct optab_pattern): Ditto.
	* optabs-query.h (optab_handler): Ditto.
	(convert_optab_handler): Ditto.