fortran: Release symbols in reversed order [PR106050]
Hello,

I saw the light regarding this PR after Paul posted a comment yesterday.

Regression test in progress on x86_64-pc-linux-gnu.
I plan to push in the next hours.

Mikael

-- >8 --

Release symbols in reversed order wrt the order they were allocated.
This fixes an error recovery ICE in the case of a misplaced
derived type declaration.  Such a declaration creates nested
symbols, one for the derived type and one for each type parameter,
which should be immediately released as the declaration is
rejected.  This breaks if the derived type is released first.
As the type parameter symbols are in the namespace of the derived
type, releasing the derived type releases the type parameters, so
one can't access them after that, even to release them.  Hence,
the type parameters should be released first.

	PR fortran/106050

gcc/fortran/ChangeLog:

	* symbol.cc (gfc_restore_last_undo_checkpoint): Release symbols
	in reverse order.

gcc/testsuite/ChangeLog:

	* gfortran.dg/pdt_33.f90: New test.