ctf: add support to pass through BTF tags
BTF generation currently relies on the internal CTF representation to
convert debug info from DWARF dies. This patch adds a new internal
header, "ctf-int.h", which defines CTF kinds to be used internally to
represent BTF tags which must pass through the CTF container. It also
adds a new type for representing information specific to those tags, and
a member for that type in struct ctf_dtdef.

This patch also updates ctf_add_reftype to accept a const char * name,
and add it for the newly added type.

gcc/

	* ctf-int.h: New file.
	* ctfc.h: Include ctf-int.h.
	(struct ctf_btf_annotation): New.
	(struct ctf_dtdef): Add dtu_btfnote member to dtd_u union.
	(enum ctf_dtu_d_union_enum): Add CTF_DTU_D_BTFNOTE.
	(ctf_add_reftype): Add name argument.
	* ctfc.cc (ctf_dtu_d_union_selector): Handle CTFC_INT_K_DECL_TAG.
	(ctf_add_reftype): Add name argument. Pass it to ctf_add_generic.
	(ctf_add_pointer): Update ctf_add_reftype call.
	* btfout.cc (btf_init_postprocess): Update ctf_add_reftype call.
	* dwarf2ctf.cc (gen_ctf_modifier_type): Update ctf_add_reftype call.