c++: non-standalone surrogate call template
Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk?  There might be an existing PR for this issue but Bugzilla search
seems to be timing out for me currently.

-- >8 --

I noticed we were accidentally preventing ourselves from considering
a pointer/reference-to-function conversion function template if it's
not the first conversion function that's considered, which for the
testcase below resulted in us accepting the B call but not the A call
despite the only difference between A and B being the order of member
declarations.  This patch fixes this so that the outcome of overload
resolution doesn't arbitrarily depend on declaration order in this
situation.

gcc/cp/ChangeLog:

	* call.cc (add_template_conv_candidate): Don't check for
	non-empty 'candidates' here.
	(build_op_call): Check it here, before we've considered any
	conversion functions.

gcc/testsuite/ChangeLog:

	* g++.dg/overload/conv-op5.C: New test.