vect: Initialize new_temp to avoid false positive warning [PR110652]
Hi,

As PR110652 and its duplicate PRs show, there could be one
build error

  error: 'new_temp' may be used uninitialized

for some build configurations.  It's a false positive warning
(or error at -Werror), but in order to make the build succeed,
this patch is to initialize the reported variable 'new_temp'
as NULL_TREE.

Confirmed this patch fixed the reported issue in PR110652
(with the same configuration).

Is it ok for trunk?

BR,
Kewen
-----
	PR tree-optimization/110652

gcc/ChangeLog:

	* tree-vect-stmts.cc (vectorizable_load): Initialize new_temp as
	NULL_TREE.