ada: Use new typedefs in gcc-interface
This changes gcc-interface to use the typedefs that were recently
introduced in gnat.  This is another step toward switching the code
generator to emit enums rather than preprocessor defines.

In a couple of spots, a 'default' case is also added.  These avoid
warnings from -Wswitch when the typedefs are changed to be enums.

gcc/ada/

	* gcc-interface/decl.cc (check_ok_for_atomic_type): Use Pragma_Id.
	* gcc-interface/trans.cc (lvalue_required_p, Pragma_to_gnu): Use
	Pragma_Id.
	(get_type_length, Attribute_to_gnu, get_atomic_access): Use
	Attribute_Id.

Tested on x86_64-pc-linux-gnu, committed on master.