c++: deducing empty type vs non-type argument pack
Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk?  Also verified by way of gcc_assert that we never see
TEMPLATE_PARM_INDEX here.

-- >8 --

Within a template parameter list, a non-type template parameter pack is
represented as a PARM_DECL.  But in a couple of spots where we need to
deduce an empty argument pack, we test for TEMPLATE_PARM_INDEX (within a
template parameter list) instead of for PARM_DECL, which means we end up
using TYPE_ARGUMENT_PACK even in the non-type case.  This patch fixes
this (seemingly harmless) bug.

gcc/cp/ChangeLog:

	* pt.cc (type_unification_real): Test for PARM_DECL instead
	of TEMPLATE_PARM_INDEX to distinguish a type vs non-type
	template parameter pack.
	(type_targs_deducible_from): Likewise.